requirements.txt
streamlit==1.54.0
crewai==1.9.3
sentence-transformers==5.2.3
faiss-cpu==1.13.2
torch==2.10.0
torchvision==0.25.0
transformers==4.46.3
pandas==2.3.3
numpy==2.4.2
Pillow==12.1.1
multidict==6.7.1
httpx==0.28.1
python-dotenv==1.1.1
orjson==3.11.7
jieba==0.42.1
